+++ anyblok/test_bloks/test_blok13/__init__.py
@@ -1,0 +1,14 @@
+# This file is a part of the AnyBlok project
+#
+#    Copyright (C) 2016 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v. 2.0. If a copy of the MPL was not distributed with this file,You can
+# obtain one at http://mozilla.org/MPL/2.0/.
+from anyblok.blok import Blok
+
+
+class TestBlok(Blok):
+
+    version = '1.0.0'
+    conflicting = ['test-blok1']